# docs/openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Thomas Anderson API",
    "version": "1.0.0",
    "description": "공동구매 플랫폼 API 문서. TypeScript + Express + Sequelize + MySQL로 구현되었습니다.",
    "contact": {
      "name": "최원빈",
      "x-student-id": "60203042"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://damara.bluerack.org",
      "description": "Production server"
    }
  ],
...
  "components": {
    "schemas": {
...
      "PostListResponse": {
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset",
          "hasNext"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Post"
            }
          },
          "total": {
            "type": "integer",
            "description": "필터 조건에 맞는 전체 게시글 수",
            "example": 42
          },
          "limit": {
            "type": "integer",
            "example": 20
          },
          "offset": {
            "type": "integer",
            "example": 0
          },
          "hasNext": {
            "type": "boolean",
            "description": "다음 페이지 존재 여부",
            "example": true
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": [
          "error",
          "message",
          "details"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "프론트 분기용 고정 에러 코드",
            "example": "POST_NOT_FOUND"
          },
          "message": {
            "type": "string",
            "description": "사용자 표시 또는 로그 확인용 에러 메시지",
            "example": "게시글을 찾을 수 없습니다."
          },
          "details": {
            "type": "object",
            "description": "필드별 오류 등 추가 세부 정보",
            "additionalProperties": true,
            "example": {}
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/posts": {
      "get": {
        "summary": "홈 피드 상품 목록 조회 (페이징, 검색, 필터, 정렬 가능)",
        "tags": [
          "Posts"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "x-user-id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "현재 사용자 UUID. 전달하면 각 게시글의 isFavorite, isParticipant, isOwner를 사용자 기준으로 계산합니다."
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "조회 개수"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "시작 위치"
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string",
              "enum": [
                "food",
                "daily",
                "beauty",
                "electronics",
                "school",
                "freemarket"
              ]
            },
            "description": "카테고리 필터 (food=먹거리, daily=일상용품, beauty=뷰티·패션, electronics=전자기기, school=학용품, freemarket=프리마켓)",
            "example": "food"
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string",
              "enum": [
                "latest",
                "deadline",
                "popular"
              ],
              "default": "latest"
            },
            "description": "정렬 방식 (latest=최신순, deadline=마감임박순, popular=인기순. popular는 currentQuantity DESC, favoriteCount DESC, createdAt DESC 기준)",
            "example": "popular"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "open",
                "closed",
                "in_progress",
                "completed",
                "cancelled"
              ]
            },
            "description": "게시글 상태 필터. 홈 화면 모집중 목록은 open을 사용합니다.",
            "example": "open"
          },
          {
            "in": "query",
            "name": "keyword",
            "schema": {
              "type": "string"
            },
            "description": "제목, 내용, 픽업 장소 검색어. q 쿼리도 같은 의미로 사용할 수 있습니다.",
            "example": "물티슈"
          },
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            },
            "description": "keyword와 같은 의미의 검색어 alias입니다.",
            "example": "물티슈"
          },
          {
            "in": "query",
            "name": "userId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "현재 사용자 UUID. x-user-id 헤더를 쓰기 어려운 클라이언트용 대체 파라미터입니다."
          }
        ],
        "responses": {
          "200": {
            "description": "상품 목록 조회 성공. items 각 항목에는 카드 UI에 필요한 관심/참여/작성자/마감 상태가 포함됩니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "상품 등록",
        "tags": [
          "Posts"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "post"
                ],
                "properties": {
                  "post": {
                    "type": "object",
                    "required": [
                      "authorId",
                      "title",
                      "content",
                      "price",
                      "minParticipants",
                      "deadline",
                      "pickupLocation"
                    ],
                    "properties": {
                      "authorId": {
                        "type": "string",
                        "format": "uuid",
                        "example": "a87522bd-bc79-47b0-a73f-46ea4068a158",
                        "description": "작성자 UUID"
                      },
                      "title": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 200,
                        "example": "맛있는 치킨 공동구매",
                        "description": "상품명"
                      },
                      "productName": {
                        "type": "string",
                        "maxLength": 200,
                        "nullable": true,
                        "example": "BBQ 황금올리브치킨 2마리 세트",
                        "description": "상세/등록 UI 상품명. 없으면 title을 fallback으로 사용합니다."
                      },
                      "content": {
                        "type": "string",
                        "minLength": 1,
                        "example": "BBQ 황금올리브치킨 2마리 세트를 함께 주문하실 분 구합니다!",
                        "description": "상품 설명"
                      },
                      "price": {
                        "type": "number",
                        "minimum": 0,
                        "example": 25000,
                        "description": "가격"
                      },
                      "minParticipants": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 2,
                        "description": "최소 참여 인원"
                      },
                      "deadline": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-27T23:59:59.000Z",
                        "description": "마감 시간 (ISO 8601 형식)"
                      },
                      "pickupLocation": {
                        "type": "string",
                        "maxLength": 200,
                        "example": "명지대학교 정문",
                        "description": "픽업 장소"
                      },
                      "pickupDate": {
                        "type": "string",
                        "format": "date",
                        "nullable": true,
                        "example": "2026-06-17",
                        "description": "수령 날짜 (YYYY-MM-DD)"
                      },
                      "pickupStartTime": {
                        "type": "string",
                        "nullable": true,
                        "example": "17:00",
                        "description": "수령 시작 시간 (HH:mm 또는 HH:mm:ss)"
                      },
                      "pickupEndTime": {
                        "type": "string",
                        "nullable": true,
                        "example": "19:00",
                        "description": "수령 종료 시간 (HH:mm 또는 HH:mm:ss)"
                      },
                      "pickupGuide": {
                        "type": "string",
                        "nullable": true,
                        "example": "정문 앞 파란 우산 근처에서 수령해 주세요.",
                        "description": "수령 안내"
                      },
                      "groupBuyType": {
                        "type": "string",
                        "nullable": true,
                        "example": "campus_pickup",
                        "description": "공구 방식/유형"
                      },
                      "tags": {
                        "type": "array",
                        "nullable": true,
                        "items": {
                          "type": "string"
                        },
                        "example": [
                          "대용량",
                          "생활용품"
                        ],
                        "description": "공구 태그 목록"
                      },
                      "notice": {
                        "type": "string",
                        "nullable": true,
                        "example": "입금 확인 후 주문 예정입니다.",
                        "description": "상세 화면 공지사항"
                      },
                      "category": {
                        "type": "string",
                        "enum": [
                          "food",
                          "daily",
                          "beauty",
                          "electronics",
                          "school",
                          "freemarket"
                        ],
                        "nullable": true,
                        "example": "food",
                        "description": "카테고리 ID (food=먹거리, daily=일상용품, beauty=뷰티·패션, electronics=전자기기, school=학용품, freemarket=프리마켓)"
                      },
                      "images": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "minLength": 1
                        },
                        "example": [
                          "https://images.unsplash.com/photo-1626082927389-6cd097cdc6ec?w=400"
                        ],
                        "description": "이미지 URL 배열 (선택사항)"
                      }
                    }
                  }
                }
              },
              "example": {
                "post": {
                  "authorId": "a87522bd-bc79-47b0-a73f-46ea4068a158",
                  "title": "맛있는 치킨 공동구매",
                  "productName": "BBQ 황금올리브치킨 2마리 세트",
                  "content": "BBQ 황금올리브치킨 2마리 세트를 함께 주문하실 분 구합니다!",
                  "price": 25000,
                  "minParticipants": 2,
                  "deadline": "2025-11-27T23:59:59.000Z",
                  "pickupLocation": "명지대학교 정문",
                  "pickupDate": "2026-06-17",
                  "pickupStartTime": "17:00",
                  "pickupEndTime": "19:00",
                  "pickupGuide": "정문 앞 파란 우산 근처에서 수령해 주세요.",
                  "groupBuyType": "campus_pickup",
                  "tags": [
                    "대용량",
                    "생활용품"
                  ],
                  "notice": "입금 확인 후 주문 예정입니다.",
                  "category": "food",
                  "images": [
                    "https://images.unsplash.com/photo-1626082927389-6cd097cdc6ec?w=400"
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "상품 등록 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "400": {
            "description": "유효성 검증 실패"
          },
          "404": {
            "description": "작성자를 찾을 수 없음"
          }
        }
      }
    },
...
    "/api/posts/{id}/status": {
      "patch": {
        "summary": "게시글 상태 변경 (작성자만 가능)",
        "tags": [
          "Posts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "게시글 UUID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status",
                  "authorId"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "open",
                      "closed",
                      "in_progress",
                      "completed",
                      "cancelled"
                    ],
                    "description": "변경할 상태",
                    "example": "closed"
                  },
                  "authorId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "작성자 UUID (권한 확인용)",
                    "example": "a87522bd-bc79-47b0-a73f-46ea4068a158"
                  }
                }
              },
              "example": {
                "status": "closed",
                "authorId": "a87522bd-bc79-47b0-a73f-46ea4068a158"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "상태 변경 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 상태 전이 또는 요청 데이터",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "권한 없음 (작성자가 아님)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "UNAUTHORIZED"
                    },
                    "message": {
                      "type": "string",
                      "example": "작성자만 상태를 변경할 수 있습니다."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "게시글을 찾을 수 없음"
          }
        }
      }
    },
...
  }
}